+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.2",
     "@uidotdev/usehooks": "^2.4.1",
-    "antd": "^5.16.0",
+    "antd": "^5.16.1",
     "class-variance-authority": "^0.7.0",
     "i18next": "^23.10.1",
     "react": "^18.2.0",
