+++ src/components/Button.tsx
@@ -9,7 +9,7 @@ const buttonStyles = cva(
     variants: {
       variant: {
         default: 'border-[1px] border-black bg-white text-black hover:opacity-80 hover:transition-all',
-        background: 'bg-primary text-white hover:opacity-90 hover:transition-all',
+        background: 'bg-primary text-white border-[1px] border-transparent hover:opacity-90 hover:transition-all',
         error: 'bg-error text-white border-none',
         link: 'text-primary underline-offset-4 hover:underline hover:transiton-all',
         outLine:
@@ -17,7 +17,7 @@ const buttonStyles = cva(
       },
       size: {
         default: 'px-5 py-2',
-        sm: 'px-6 py-3 rounded-md',
+        md: 'px-6 py-3 rounded-md',
         lg: 'px-7 py-3.5 rounded-mb text-lg',
         icon: 'h-10 w-10'
       }
