+++ src/locales/en/auth.json
@@ -1,9 +1,9 @@
 {
   "auth": {
-    "signupHeading": "Sign Up for an Account",
+    "signupHeading": "Get’s started.",
     "signupSub": "Already have an account?",
-    "signinHeading": "Sign In to Your Account",
-    "signinSub": "Don't have an account yet?",
+    "signinHeading": "Get’s started.",
+    "signinSub": "Don’t have an account?",
     "signin": "Sign In",
     "signup": "Sign Up",
     "email": "Email",
