+++ src/locales/en/home.json
@@ -6,8 +6,8 @@
       "study": "Study",
       "new": "News",
       "more": "Read News",
-      "register": "Register",
-      "login": "Log In"
+      "register": "Sign Up",
+      "login": "Sign In"
     },
     "language": {
       "vi": "Vietnamese",
