+++ src/locales/vi/home.json
@@ -14,7 +14,7 @@
       "en": "Tiếng Anh"
     },
     "banner": {
-      "heading": "Khám phá những điều tốt nhất trong Giải trí hôm nay",
+      "heading": "Khám phá những trải nghiệm giải trí đỉnh cao hôm nay.",
       "subHeading": "Khám phá một loạt các bài hát, phim, tin tức và nhiều hơn nữa.",
       "learnMore": "Tìm hiểu thêm"
     },
