+++ src/main.tsx
@@ -1,9 +1,9 @@
+import './index.css'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
-import { store } from '~/app/store.ts'
+import { store } from '~/redux/store.ts'
 import App from './App.tsx'
-import './index.css'
 import '~/i18n/i18n.ts'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
